move seeds script to lib/auth/init for #174
