+++ .env_sample
@@ -1,9 +1,15 @@
 export ADMIN_EMAIL=youremail@gmail.com
-export AUTH_API_KEY=2PzB7PPnpuLsbWmWtXpGyI+kfSQSQ1zUW2Atz/+8PdZuSEJzHgzGnJWV35nTKRwx/dwylauth.herokuapp.com
-export EMAIL_APP_URL=https://dwylmail.herokuapp.com
+export AUTH_URL=dwylauth.herokuapp.com
+export SECRET_KEY_BASE=2PzB7PPnpuLsbWmWtXpGyI+kfSQSQ1zUW2Atz/+8PdZuSEJzHgzGnJWV35nTKRwx
+export ENCRYPTION_KEYS='nMdayQpR0aoasLaq1g94FLba+A+wB44JLko47sVQXMg=,L+ZVX8iheoqgqb22mUpATmMDsvVGtafoAeb0KN5uWf0='
+
 export GITHUB_CLIENT_ID=CreateGitHubApp
 export GITHUB_CLIENT_SECRET=SuperSecret
 export GOOGLE_CLIENT_ID=YourAppsClientId.apps.googleusercontent.com
 export GOOGLE_CLIENT_SECRET=SuperSecret
-export SECRET_KEY_BASE=2PzB7PPnpuLsbWmWtXpGyI+kfSQSQ1zUW2Atz/+8PdZuSEJzHgzGnJWV35nTKRwx
-export ENCRYPTION_KEYS='nMdayQpR0aoasLaq1g94FLba+A+wB44JLko47sVQXMg=,L+ZVX8iheoqgqb22mUpATmMDsvVGtafoAeb0KN5uWf0='
+
+# Optional for sending emails:
+export EMAIL_APP_URL=https://dwylmail.herokuapp.com
+
+# Export this once your the auth db is initialized
+export AUTH_API_KEY=2PzB7PPnpuLsbWmWtXpGyI+kfSQSQ1zUW2Atz/+8PdZuSEJzHgzGnJWV35nTKRwx/dwylauth.herokuapp.com
