+++ mix.exs
@@ -70,8 +70,11 @@ defmodule Auth.Mixfile do
       # https://github.com/dwyl/elixir-auth-google
       {:elixir_auth_google, "~> 1.6.1"},
 
+      # Check/get Environment Variables: https://github.com/dwyl/envar
+      {:envar, "~> 1.0.4"},
+
       # https://github.com/dwyl/auth_plug
-      {:auth_plug, "~> 1.4.7"},
+      {:auth_plug, "~> 1.4.11 "},
 
       # https://github.com/dwyl/rbac
       {:rbac, "~> 0.5.3"},
