+++ test/auth/init/init_test.exs
@@ -1,0 +1,10 @@
+defmodule Auth.InitTest do
+  use Auth.DataCase, async: true
+
+  describe "Initialize the Auth Database" do
+    test "main/0" do
+      assert Auth.Init.main() == :ok
+    end
+  end
+  
+end
